% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.cov.R
\name{pre.cov}
\alias{pre.cov}
\title{Generate LD files}
\usage{
pre.cov(ldpath = NULL, bedfile, path, chr = 1:22, shrink = T)
}
\arguments{
\item{ldpath}{path for plink bfiles for the chromosome-wise LD reference;(e.g. 1000 Genome Project)}

\item{bedfile}{path to the bedfile for OWAS segments (default: 100 KB up and down-stream of the TSS, the length of each segment is 5 KB)}

\item{path}{specify the output path}

\item{chr}{chromosome, can be a integer or vector from 1 to 22, default=1:22}

\item{shrink}{shrink LD matrix or not, default=T}
}
\description{
Using this function to generate blockwise LD matrix
}
