% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.sc.R
\name{pre.sc}
\alias{pre.sc}
\title{Generate openness score files}
\usage{
pre.sc(ctype, scpath, ldpath = NULL, path, chr = 1:22)
}
\arguments{
\item{ctype}{celltype}

\item{scpath}{path to the openness scores files}

\item{ldpath}{path for plink bfiles for the chromosome-wise LD reference;(e.g. 1000 Genome Project)}

\item{path}{specify the output path}

\item{chr}{chromosome, can be a integer or vector from 1 to 22, default=1:22}
}
\description{
Using this function to generate blockwise openness scores
}
