% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.owas.R
\name{run.owas}
\alias{run.owas}
\title{Main function}
\usage{
run.owas(
  ctype,
  gwas,
  trait = "test",
  ldpath,
  path,
  plinkpath,
  chr = 1:22,
  clump = T
)
}
\arguments{
\item{ctype}{celltype}

\item{gwas}{GWAS summary statistics, need to include rsid, chr, a1, a2, z, p (header is necessary)}

\item{trait}{trait name}

\item{ldpath}{path for plink bfiles for the chromosome-wise LD reference;(e.g. 1000 Genome Project)}

\item{path}{specify the output path}

\item{plinkpath}{chromosome, can be a integer or vector from 1 to 22, default=1:22}

\item{chr}{chromosome, can be a integer or vector from 1 to 22, default=1:22}
}
\description{
Run OWAS
}
